[feat] QuestRecord, Quest의 entity, controller, service

+++ src/main/java/com/likelion/MoodMate/service/UserService.java
@@ -5,7 +5,6 @@ import com.likelion.MoodMate.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -14,19 +13,32 @@ public class UserService {
     @Autowired
     private UserRepository userRepository;
 
-    public List<User> findAll() {
-        return userRepository.findAll();
-    }
-
-    public Optional<User> findById(Long id) {
-        return userRepository.findById(id);
-    }
-
-    public User save(User user) {
+    public User createUser(User user) {
         return userRepository.save(user);
     }
 
-    public void deleteById(Long id) {
-        userRepository.deleteById(id);
+    public Optional<User> getUser(Long id) {
+        return userRepository.findById(id);
+    }
+
+    public Optional<User> updateUser(Long id, User userDetails) {
+        Optional<User> optionalUser = userRepository.findById(id);
+        if (optionalUser.isPresent()) {
+            User user = optionalUser.get();
+            user.setUserName(userDetails.getUserName());
+            user.setUserPassword(userDetails.getUserPassword());
+            return Optional.of(userRepository.save(user));
+        } else {
+            return Optional.empty();
+        }
+    }
+
+    public boolean deleteUser(Long id) {
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 }
